Add more lockAndMockStatic
