+++ core/src/test/kotlin/org/stellar/anchor/util/LogTest.kt
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.TestInstance
 import org.slf4j.Logger
 import org.stellar.anchor.config.AppConfig
 import org.stellar.anchor.config.PII
+import org.stellar.anchor.lockAndMockStatic
 import org.stellar.anchor.util.Log.shorter
 import org.stellar.sdk.Network.TESTNET
 
@@ -22,15 +23,9 @@ internal class LogTest {
   @BeforeEach
   fun setup() {
     MockKAnnotations.init(this, relaxed = true)
-    mockkStatic(Log::class)
-    every { Log.getLogger() } returns logger
   }
 
-  @AfterEach
-  fun teardown() {
-    clearAllMocks()
-    unmockkAll()
-  }
+  @AfterEach fun teardown() {}
 
   @Suppress("unused")
   class TestBean {
@@ -49,40 +44,47 @@ internal class LogTest {
 
   @Test
   fun `test log messages`() {
-    Log.error("Hello")
-    verify { logger.error("Hello") }
+    lockAndMockStatic(Log::class) {
+      every { Log.getLogger() } returns logger
 
-    Log.warn("Hello")
-    verify { logger.warn("Hello") }
+      Log.error("Hello")
+      verify { logger.error("Hello") }
 
-    Log.info("Hello")
-    verify { logger.info("Hello") }
+      Log.warn("Hello")
+      verify { logger.warn("Hello") }
 
-    Log.debug("Hello")
-    verify { logger.debug("Hello") }
+      Log.info("Hello")
+      verify { logger.info("Hello") }
 
-    Log.trace("Hello")
-    verify { logger.trace("Hello") }
+      Log.debug("Hello")
+      verify { logger.debug("Hello") }
+
+      Log.trace("Hello")
+      verify { logger.trace("Hello") }
+    }
   }
 
   @Test
   fun `test log messages with JSON format`() {
-    val detail = TestBeanPII()
+    lockAndMockStatic(Log::class) {
+      every { Log.getLogger() } returns logger
+      val detail = TestBeanPII()
 
-    Log.error("Hello", detail)
-    verify { logger.error("Hello$wantTestPIIJson") }
+      Log.error("Hello", detail)
+      verify { logger.error("Hello$wantTestPIIJson") }
 
-    Log.warn("Hello", detail)
-    verify { logger.warn("Hello$wantTestPIIJson") }
+      Log.warn("Hello", detail)
+      verify { logger.warn("Hello$wantTestPIIJson") }
 
-    Log.info("Hello", detail)
-    verify { logger.info("Hello$wantTestPIIJson") }
+      Log.info("Hello", detail)
+      verify { logger.info("Hello$wantTestPIIJson") }
 
-    Log.debug("Hello", detail)
-    verify { logger.debug("Hello$wantTestPIIJson") }
+      Log.debug("Hello", detail)
+      verify { logger.debug("Hello$wantTestPIIJson") }
 
-    Log.trace("Hello", detail)
-    verify { logger.trace("Hello$wantTestPIIJson") }
+      Log.trace("Hello", detail)
+      verify { logger.trace("Hello$wantTestPIIJson") }
+    }
   }
 
   @Suppress("unused")
@@ -106,15 +108,18 @@ internal class LogTest {
 
   @Test
   fun `test errorEx`() {
-    Log.errorEx(Exception("mock exception"))
-    verify(exactly = 1) { logger.error(any()) }
+    lockAndMockStatic(Log::class) {
+      every { Log.getLogger() } returns logger
+      Log.errorEx(Exception("mock exception"))
+      verify(exactly = 1) { logger.error(any()) }
 
-    val slot = slot<String>()
-    every { logger.error(capture(slot)) } answers {}
+      val slot = slot<String>()
+      every { logger.error(capture(slot)) } answers {}
 
-    Log.errorEx("Hello", Exception("mock exception"))
-    assertTrue(slot.captured.contains("Hello"))
-    assertTrue(slot.captured.contains("mock exception"))
+      Log.errorEx("Hello", Exception("mock exception"))
+      assertTrue(slot.captured.contains("Hello"))
+      assertTrue(slot.captured.contains("mock exception"))
+    }
   }
 
   @Test
@@ -129,7 +134,6 @@ internal class LogTest {
 
   @Test
   fun `test getLogger`() {
-    unmockkAll()
     val logger = Log.getLogger()
     assertNotNull(logger)
   }
